Apply the remaining suggestions
